split item updating and impl specific code
