+++ items/src/main/scala/weco/api/items/services/ItemUpdateService.scala
@@ -1,75 +1,21 @@
 package weco.api.items.services
 
-import grizzled.slf4j.Logging
-import weco.api.stacks.models.SierraItemIdentifier
-import weco.api.stacks.services.SierraService
 import weco.catalogue.internal_model.identifiers.{IdState, IdentifierType, SourceIdentifier}
-import weco.catalogue.internal_model.locations.{AccessCondition, PhysicalLocation}
 import weco.catalogue.internal_model.work.{Item, Work, WorkState}
-import weco.catalogue.source_model.sierra.identifiers.SierraItemNumber
-
-import scala.collection.immutable
 import scala.concurrent.{ExecutionContext, Future}
 
 trait ItemUpdater {
+  val identifierType: IdentifierType
   def updateItems(items: Seq[Item[IdState.Minted]]): Future[Seq[Item[IdState.Minted]]]
 }
 
-class SierraItemUpdater(sierraService: SierraService)(implicit executionContext: ExecutionContext)
-  extends ItemUpdater
-    with Logging {
-
-  private def updateLocations(
-                               item: Item[IdState.Minted],
-                               accessCondition: AccessCondition
-                             ) = item.locations.map {
-    case physicalLocation: PhysicalLocation =>
-      physicalLocation.copy(
-        accessConditions = List(accessCondition)
-      )
-    case location => location
-  }
-
-  private def updateAccessConditions(
-                                      sierraItemSourceIdentifiers: Map[SierraItemNumber, Item[IdState.Minted]],
-                                      accessConditions: Map[SierraItemNumber, Option[AccessCondition]]
-                                    ): immutable.Iterable[Item[IdState.Minted]] = {
-    accessConditions.flatMap {
-      case (itemNumber, Some(accessCondition)) => {
-        sierraItemSourceIdentifiers
-          .get(itemNumber)
-          .map(item => {
-            item.copy(
-              locations = updateLocations(item, accessCondition)
-            )
-          })
-      }
-    }
-  }
-
-  def updateItems(items: Seq[Item[IdState.Minted]]) = {
-    val sierraItemSourceIdentifiers = items.map {
-      case item@Item(IdState.Identified(_, srcId, _), _, _, _) =>
-        SierraItemIdentifier.fromSourceIdentifier(srcId) -> item
-    } toMap
-
-    val itemNumbers = sierraItemSourceIdentifiers.keys.toSeq
-
-    sierraService.getAccessConditions(itemNumbers)
-      .map {
-        case Right(accessConditions) =>
-          updateAccessConditions(sierraItemSourceIdentifiers, accessConditions)
-        case Left(err) =>
-          error(msg = f"Couldn't refresh items for $itemNumbers got error $err")
-          items
-      } map(_.toSeq)
-  }
-}
-
-
 class ItemUpdateService(
-  sierraItemUpdater: SierraItemUpdater
+  itemUpdaters: List[ItemUpdater]
 )(implicit executionContext: ExecutionContext) {
+
+  private final val itemUpdatesMap = itemUpdaters
+    .map(updater => updater.identifierType -> updater)
+    .toMap
 
   private def buildItemsMap(items: Seq[Item[IdState.Minted]]): Map[SourceIdentifier, Item[IdState.Minted]] =
     items.map {
@@ -107,10 +53,10 @@ class ItemUpdateService(
       work.data.items.zipWithIndex.groupBy {
         case (Item(IdState.Identified(_, srcId, _), _, _, _), _) => srcId.identifierType
       } map {
-        case (IdentifierType.SierraSystemNumber, itemsWithIndex) =>
-          preserveOrder(itemsWithIndex, sierraItemUpdater.updateItems)
-        // In the future if we wish to update other sources we can add that here
-        case (_, itemsWithIndex) => Future(itemsWithIndex)
+        case (identifierType, itemsWithIndex) => itemUpdatesMap
+          .get(identifierType)
+          .map(updater => preserveOrder(itemsWithIndex, updater.updateItems))
+          .getOrElse(Future(itemsWithIndex))
       }
       // unzipWithIndex
     } map(_.flatten.toList.sortBy(_._2).map(_._1))
